+++ docker/makebusy/run.sh
@@ -1,19 +1,25 @@
 #!/bin/bash
 
-if [ ! -z $TESTS_PATH ] ; then
+NETWORK=${NETWORK:-"kazoo"}
+NAME=makebusy.$NETWORK
+
+if [ ! -z $TESTS_PATH ]
+then
 	VOLUME="-v $TESTS_PATH:/home/user/make-busy/tests/KazooTests/Applications/"
 else
 	echo Probably useless, please specify where are your tests in TESTS_PATH env variable
 	VOLUME=""
 fi
 
-NETWORK=${NETWORK:-"kazoo"}
-NAME=makebusy.$NETWORK
+if [ -n "$(docker ps -aq -f name=$NAME)" ]
+then
+   echo -n "stopping: "
+   docker stop -t 1 $NAME
+   echo -n "removing: "
+   docker rm -f $NAME
+fi
 
-echo :: starting $NAME tests:$TESTS_PATH
-
-docker stop -t 1 $NAME
-docker rm -f $NAME
+echo -n "starting: $NAME tests: $TESTS_PATH"
 docker run -td \
 	--net $NETWORK \
 	-h $NAME \
